move backend creation to profile context
keeping it in the runner will create conflicts with rspec runners
