+++ lib/vulcano/profile_context.rb
@@ -4,6 +4,7 @@
 
 require 'vulcano/rule'
 require 'vulcano/dsl'
+require 'train'
 
 module Vulcano
   class ProfileContext
@@ -139,6 +140,37 @@ module Vulcano
         end
       end
       # rubocop:enable all
+    end
+
+    # Create the transport backend with aggregated resources.
+    #
+    # @param [Hash] config for the transport backend
+    # @return [Transport] enriched transport instance
+    def self.create_backend(config)
+      conf = Train.target_config(config)
+      name = conf[:backend] || :local
+      transport = Train.create(name, conf)
+      if transport.nil?
+        fail "Can't find transport backend '#{name}'."
+      end
+
+      connection = transport.connection
+      if connection.nil?
+        fail "Can't connect to transport backend '#{name}'."
+      end
+
+      cls = Class.new do
+        define_method :backend do
+          connection
+        end
+        Vulcano::Resource.registry.each do |id, r|
+          define_method id.to_sym do |*args|
+            r.new(self, id.to_s, *args)
+          end
+        end
+      end
+
+      cls.new
     end
   end
 end
