+++ test/helper.rb
@@ -14,7 +14,7 @@ SimpleCov.start do
 end
 
 require 'vulcano/resource'
-require 'train'
+require 'vulcano/profile_context'
 
 class MockLoader
   # pass the os identifier to emulate a specific operating system
@@ -47,8 +47,11 @@ class MockLoader
     scriptpath = ::File.realpath(::File.dirname(__FILE__))
 
     # create mock backend
-    @backend = Train.create('mock')
-    mock = @backend.connection
+    @backend = Vulcano::ProfileContext.create_backend({ backend: :mock })
+    mock = @backend.backend
+
+    # set os emulation
+    mock.mock_os(@os)
 
     # create all mock files
     local = Train.create('local').connection
@@ -92,11 +95,11 @@ class MockLoader
     # create all mock commands
     cmd = lambda {|x|
       stdout = ::File.read(::File.join(scriptpath, '/unit/mock/cmd/'+x))
-      mock.mock_command(stdout, '', 0)
+      mock.mock_command('', stdout, '', 0)
     }
 
     empty = lambda {
-      mock.mock_command('', '', 0)
+      mock.mock_command('', '', '', 0)
     }
 
     mock.commands = {
@@ -178,9 +181,6 @@ class MockLoader
       # iptables
       'iptables  -S' => cmd.call('iptables-s'),
     }
-
-    # set os emulation
-    mock.os = @os
 
     @backend
   end
